fix(import): address upload handler review findings
Filename sanitisation before DB insert; InlineRouter replaced by TestRouter.post_multipart.
mod.rs ordering fixed; i64 row_count; distinct error messages; dead lint allow removed.

+++ backend/src/dto/import.rs
@@ -90,7 +90,7 @@ pub struct UploadResponse {
     pub headers: Vec<String>,
     #[schema(value_type = Vec<Object>)]
     pub preview: Vec<BTreeMap<String, Option<String>>>,
-    pub row_count: i32,
+    pub row_count: i64,
 }
 
 // ── Request DTOs ───────────────────────────────────────────────────────────────
